ボタン部分を共通化
